My solutions
