--- Data_Structure/requirements.txt
@@ -1,4 +1,0 @@
-
-pytest
-mypy
-flake8
